sync r_amlogic_t7_20221021_64_bit_rc2.xml code >>>>>>
commit ae1b9237740cb258cd032803f5e439df6366850b (HEAD, openlinux/openlinux/ds/r-amlogic-64-bit-20220111, m/r-amlogic)
Author: Yunmin Chen <yunmin.chen@amlogic.com>
Date:   Thu Jul 21 14:03:38 2022 +0800

    mediahal: CB2 disable 64bit library compilation [1/1]

    PD#SWPL-83263

    Problem:
    S/T disable 64bit library compilation

    Solution:
    S/T disable 64bit library compilation

    Verify:
    SC2

    Build Tag:
    Android_R_DRM_PATCHBUILD-6208

    Commits:
    media_hal:
    version: 3.10.29-g0ff07bc(build:6208)
    0ff07bc mediahal: CB2 disable 64bit library compilation [1/1]

    hardware/amlogic/media_modules:
    d8a7a814 v4l: CB2 fix video playback stuck and mosaic. [1/1]

    vendor/amlogic/common:
    d35ac0db0 AmNuPlayer: CB2 Switch the audio track screen stuck for more

    hardware/amlogic/audio:
    c250741c audio: fixed dvb audio easing vol issue [1/1]

    hardware/amlogic/gralloc:
    73090f4 gralloc: fix camera usage buffer allocation failure [1/1]

    Git Hash-Object:
    564af08718e583672d6134b1eb7e1d10a6ce4faf
    Android.bp
    fc389f1048d5ffa0365ee055644d8841f5a3133c
    arm64/bin/AmTsPlayerExample
.....

+++ example/EsVideoDecPlayer/EsVideoDecPlayer.cpp
@@ -33,9 +33,8 @@
 #include <getopt.h>
 #include <thread>
 #include <sys/mman.h>
-#include "crc32.h"
 #include "AmVideoDecBase.h"
-
+#include <inttypes.h>
 static uint32_t kDefaultQueueCount = 32;
 using namespace std::chrono_literals;
 #define MAX_INSTANCE_MUN  9
@@ -314,13 +313,14 @@ void VideoDecPlayerExample::dumpData(char* buf, uint32_t width, uint32_t height,
     if (mcFp) {
         /* save output crc data */
         int crc_y = 0,crc_uv = 0;
-
+        /*
         for (i = 0 ; i < (int)height; i++) {
             crc_y = crc32_le(crc_y, (unsigned char const *)(buf + i * mBufferWidth), width);
         }
         for (i = 0 ; i < (int)height / 2; i++) {
             crc_uv = crc32_le(crc_uv, (unsigned char const *)(buf + y_size + i * mBufferWidth), width);
         }
+        */
         fprintf(mcFp, "%08d: %08x %08x\n", mDumpNum, crc_y, crc_uv);
         fflush(mcFp);
         fsync(fileno(mcFp));
@@ -613,17 +613,17 @@ void VideoDecPlayerExample::onOutputFormatChanged(uint32_t requestedNumOfBuffers
     }
 #endif
 
-    printf("onOutputFormatChanged out timeUs %lld\n",(long long)getTimeUsFromStart());
+    printf("onOutputFormatChanged out timeUs %" PRId64 "\n", getTimeUsFromStart());
 }
 
 
 void VideoDecPlayerExample::onOutputBufferDone(int32_t pictureBufferId, int64_t bitstreamId,
                                                uint32_t width, uint32_t height) {
-    printf("onOutputBufferDone this %p, pictureBufferId %d, bitstreamId %lld, mInputWork.size %d, In %d, Out %d\n",
-                        this, pictureBufferId,(long long)bitstreamId,(int)mInputWork.size(), mInputDoneCount, mOutputDoneCount);
+    printf("onOutputBufferDone this %p, pictureBufferId %d, bitstreamId %" PRId64 ", mInputWork.size %d, In %d, Out %d\n",
+                        this, pictureBufferId, bitstreamId, (int)mInputWork.size(), mInputDoneCount, mOutputDoneCount);
     std::lock_guard<std::mutex> lock(mOutputLock);
     int64_t timestamp = mInputWork[bitstreamId];
-    printf("onOutputBufferDone this %p, timestamp %lld, width %d, height %d\n",this,(long long)timestamp, width, height);
+    printf("onOutputBufferDone this %p, timestamp %" PRId64 ", width %d, height %d\n",this, timestamp, width, height);
 
     mDisplayWork.push({bitstreamId, timestamp, pictureBufferId, width, height});
     mInputWork.erase(bitstreamId);
